Moved the composable to a different file and added icons
