+++ app/src/main/res/drawable/hide_password.xml
@@ -1,0 +1,27 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="25dp"
+    android:height="25dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M2,2L22,22"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M6.713,6.723C3.665,8.795 2,12 2,12C2,12 5.636,19 12,19C14.05,19 15.817,18.273 17.271,17.288M11,5.058C11.325,5.02 11.659,5 12,5C18.364,5 22,12 22,12C22,12 21.308,13.332 20,14.833"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M14,14.236C13.469,14.711 12.768,15 12,15C10.343,15 9,13.657 9,12C9,11.176 9.332,10.43 9.869,9.888"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+</vector>
